improvement: make Ash.Query the source of truth for pagination
Read actions still accept a :page option, but this (like, e.g., the :tenant
option) is used to populate the :page field of the query. Moreover, it is
subsequently deleted from options, so that opts can be safely stored in the
rerun field of pages without having two different sources of truth on
pagination.